fixed warning
git-svn-id: svn://scm.gforge.inria.fr/svnroot/ecm/trunk@530 404564d9-a503-0410-82bf-e18ce2cf3989

+++ auxi.c
@@ -113,7 +113,7 @@ xmalloc (size_t size)
   p = malloc (size);
   if (p == NULL)
     {
-      fprintf (stderr, "Could not allocate %d bytes\n", size);
+      fprintf (stderr, "Could not allocate %lu bytes\n", (unsigned long) size);
       exit (EXIT_FAILURE);
     }
